refactor merkle_tree
