+++ src/merkle_tree/bitcoin_script/mod.rs
@@ -1,0 +1,2 @@
+mod twin_proof;
+pub use twin_proof::*;
